+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Allow `github.com/VictoriaMetrics/metricsql` module in SLO plugins.
 - Contrib plugin: `sloth.dev/contrib/validate_victoria_metrics/v1`.
 - Contrib plugin: `sloth.dev/contrib/rule_intervals/v1`.
+- Contrib plugin: `sloth.dev/contrib/error_budget_exhausted_alert/v1`.
+- Contrib plugin: `sloth.dev/contrib/denominator_corrected_rules/v1`.
 
 ## [v0.13.0] - 2025-09-10